feat(core): renamed domain package into model
