+++ core/src/main/java/com/example/sto/core/kafka/StatusChangeListener.java
@@ -1,7 +1,7 @@
 package com.example.sto.core.kafka;
 
 import com.example.sto.common.domain.RequestStatus;
-import com.example.sto.core.history.domain.RequestStatusHistory;
+import com.example.sto.core.history.model.RequestStatusHistory;
 import com.example.sto.core.history.repository.RequestStatusHistoryRepository;
 import com.example.sto.core.request.repository.RequestRepository;
 import com.example.sto.notification.Notifier;
