+++ core/src/main/java/com/example/sto/core/request/dto/RequestDto.java
@@ -1,7 +1,7 @@
 package com.example.sto.core.request.dto;
 
 import com.example.sto.common.domain.RequestStatus;
-import com.example.sto.core.request.domain.Request;
+import com.example.sto.core.request.model.Request;
 import lombok.AllArgsConstructor;
 import lombok.Data;
