+++ core/src/main/java/com/example/sto/core/request/service/RequestService.java
@@ -1,7 +1,7 @@
 package com.example.sto.core.request.service;
 
 import com.example.sto.common.domain.RequestStatus;
-import com.example.sto.core.request.domain.Request;
+import com.example.sto.core.request.model.Request;
 
 import java.util.List;
